_archive/pm-tools_archived_20251225/v2/backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.10.0
pydantic-settings>=2.6.0
pillow>=11.0.0
python-multipart>=0.0.17
